Implement atlasing in image.c, use for tilesets
This should improve map rendering performance.  Not that it was terrible
before, but...

Atlasing is done via a subimage system.  You create a subimage from a
section of the atlas and from there it acts like any other image.  The
best part?  Images are refcounted, so you don't have to keep a pointer
to the atlas around--you can free it immediately, and once the last
subimage is freed, the atlas is automatically freed too!