Clean up extensions list
`sphere-legacy-api` more clearly conveys what the extension represents;
`sphere-legacy` was ambiguous at best and offputting at worst--as it
suggested the engine was outdated.

I'm thinking the standard for GetExtensions() will be:
`sphere-legacy-api` if it's implemented; the engine name if the engine
includes any non-standard functionality; then any standardized
extensions in alphabetical order.