Fix edge script activation, again
So um, Sphere still runs edge scripts even if the camera isn't actually
attached to anyone, just so long as the camera itself is over an edge.
Which is nice--this allows custom camera logic without breaking edge
script activation.  I will say, for as buggy an engine as the original
Sphere was, it did cover all the bases!