Actually fix sound deadlocks this time
Well... I'm about 90% sure this fixes it, anyway.  I did a bunch of
tests with and without this change, and I haven't seen it lock up yet
with it included.  What it looks like is that you have to call
`al_drain_audio_stream` on a stream before destroying it to ensure the
feeder doesn't get stuck waiting for a fragment event that will never
come.