Refactor script handling, fixing bugs
script.c has been completely refactored, fixing several bugs related to
using functions as person/map scripts.  Specifically, using the same
function for multiple scripts was completely broken due to the way
script-in-use tracking was handled.

This wasn't originally supposed to be a refactoring exercise, but it
needed to be done anyway.  The script handling code was quite old and
fragile.