Reassign leader when person is destroyed
Full Sphere compliance requires that, when a person is destroyed, all
persons following him are made independent by having their leader set to
NULL.  For a recursive follower chain, this severs the chain, leaving
all persons behind the one removed inert and useless.  It seems
generally more useful to simply condense the chain in this case.