Fix `instanceof` for Color objects
Duktape doesn't create a prototype for native-side functions by default.
This tripped me up with the Color type, as I didn't define any methods
on it, and thus the prototype was never created.  This in turn prevented
`instanceof` from working with Color objects.

So that this won't happen again, there's now a dedicated function in
api.c for registering constructors.