Implement vector object
Wow, this is levels of type-unsafe that even I'M uncomfortable with.
And this is coming from an accomplished C programmer!  That's probably
not a good sign...

I'm going to commit this in an experimental branch for now until I'm
confident it won't lead to a thousand annoying heisenbugs down the road.