Re-centralize extension strings
This seems more maintainable to me.  Particularly in the case of Galileo
where multiple object types are defined in one file, if I were to decide
to split them, which one would get the responsibility of registering the
extension?

Everything in moderation, as they say, and that goes for encapsulation
as well.  Pragmatism is a wonderful virtue.