Codify logging verbosity levels
Level 0 shows a very small subset of initialization and shutdown
entries; this is mainly meant to silence the engine completely to ensure
you only see game output.

Level 1, the default, includes full startup and shutdown logging but is
mostly silent while in-game.

Level 2 logs when assets are loaded from disk.

Level 3 logs an entry, e.g., any time an image or surface is created.
Loading of compound assets such as tilesets or spritesets tends to be
very noisy.

Level 4 is the most verbose, logging such technical aspects as pixel
cache management for images, etc.  This is not generally practical
unless you are redirecting the output to a file.