Clean up get_asset_path() and variant
These are some of the oldest functions in the engine and haven't been
touched in a long, long time.  As they are used pervasively throughout
the engine, this is a good thing--it means they are very stable!--but
unfortunately they were a mess and didn't match the rest of the
project's coding style at all.