Implement Galileo-based GradientCircle
This is a proof-of-concept and needs quite a bit of work before it sees
production use.  As it creates and destroys a Shape and Group for every
single circle (not to mention uploading a bunch of vertices), it's
horrendously inefficient.  But the point is, it works!  This marks the
first internal use of Galileo. :smile: