Draw GradientCircles in immediate mode
I could memoize Groups for drawn circles, but that's adding unnecessary
extra complexity.  Better to draw in immediate mode and let Allegro
handle the details.  If someone wants to get the full power out of the
graphics hardware, they should be using the Galileo API directly, not
the classic functions.