Force trigger check if input person changes
Even this isn't technically fully compatible with Sphere--which tracks
trigger activation per-person--but it does fix triggers in Kefka's
Revenge, so I'll take it. :-)

I also fixed a potential crash when changing maps in a delay script and
moved update script execution after delay scripts are processed for
better compatibility with Sphere.