Fix handling of NUL terminator in word wrap
NULs were incorrectly being factored into carry determination, sometimes
causing the last word to be lost.  I wish I had caught this before the
1.2 release, but it's a pretty specific set of circumstances that cause
it--the terminating NUL has to occur at a wrap boundary.  The current
word is then carried, the loop exits having seen the NUL, and as a
result, the new line isn't committed and the word is lost.