Make lstring_t Unicode aware
A new function, lstr_make_utf8(), converts the lstring to UTF-8 if it
isn't already.  Besides making the rest of the engine more
Unicode-friendly, this has the nice bonus that the Duktape compiler
never sees an invalid UTF-8 stream.  As a result I was about to yank out
some of the fallback logic from script.c.