Update version in Git to 1.6.0
I think I'm going to make this the policy from now on, the version
number in master will be that of the planned next version, as opposed to
the previous release.  This will make it easier to distinguish
work-in-progress releases.

To prevent automated tools (such as Inno Setup) from ever treating a Git
build as the same as (or newer than) an official release, the build
number will be set to zero.