Fix Duktape error intercept
This had several bugs due to limitations in the way Duktape's debugger
is wired under the hood.  The solution to get it to work in all cases is
a bit... barbaric, but it seems to work stably in any case.  Hopefully
it doesn't cause something stupid to happen like a pig eating all the
minisphere code...