Switch to Duktape 1.3.0 official
This disables error intercept while debugging, but honestly that feature
proved a bit useless with the current Duktape debugger.  I've found that
in practice, by the time an error is thrown, the offender is several
levels down the callstack.  Since Duktape only allows you to view locals
from the *current* activation, pausing at the point of throw is of
questionable utility.