Simplify code for path resolver
The path is now replaced in-place using construct_path(). This gets rid
of the ugly boilerplate to free the existing strings, etc. and moves it
into construct_path() where it can also be used by other functions.

The relativizer is no longer part of the public API due to the stringent
requirements for its input--it doesn't access the file system and
requires a canonical path to work properly.  If an application needs a
relative path, it can get it from path_resolve().