Perform dry run on libpng integration in Cell
So far so good.  I'm not a big fan of the PNG API style, but it
definitely gets the job done, and does at least tend to get out of its
own way once you've gotten the hang of it.  minisphere's MNG support is
a testament to that: It hasn't really changed at all since I originally
wrote it!