Incorporate source map support into minisphere
Previously, source maps were handled by minisphere Remote, outside of
the engine.  This doesn't work for SPK-packaged games, however, where
the debugger doesn't have the means to read the file, and is potentially
more error-prone.  Better to let the engine do the translation, it has
more information to work with anyway.