Avoid use of undocumented Duktape function
It turns out I never needed to call duk_error_va_raw() in the first
place--simply pushing an error object and modifying its fileName and
lineNumber properties after the fact is sufficient.

Better this way than to rely on undocumented functions that are prone to
removal at any time (in theory at least).