Update to latest self-built Allegro
It turns out using the official Allegro binaries actually increased the
size of the engine by nearly 500KB, quite a hefty increase, largely due
to the inclusion of a Direct3D backend not needed by minisphere.  This
brings it back to where it was, at ~2.1MB for x86, ~2.4MB for x64.