Implement fallback for lack of shader support
Graphics cards prior to OpenGL 2.0 don't support a programmable pipeline
for shaders.  This change allows minisphere to start up with shader
support disabled.

Not sure how useful this change is in the real world, but it seemed easy
enough to implement the fallback and many games won't need shaders
anyway, so why not?