SSJ: Add some more dvalue parsing
I'm thinking the best approach will ultimately be to split the Duktape
client code into two layers: The base dvalue protocol, which sends and
receives dvalues, with the actual Duktape command processing as a
separate component.

The rationale for this is that the C# client conflates these two layers,
and code complexity suffers for it.  DuktapeClient.cs isn't exactly fun
to maintain.