Implement barebones command line
In the process of doing this I was blown away that the Windows console
seems to offer full line editing functionality (history, etc.) even for
a C app using lowly getchar().  As much of a pain as it is to develop on
Windows, I have to admit userland programs seem to have it easy.