Implement first proof-of-concept of SSJ
It's alive!  This came together much faster than I was expecting given
how tedious it is working with buffers in C.  Moving message assembly
into remote.c turned out to be a smart decision.  Now if only I could
get rid of all the manual micromanagement of message pointers, we'd be
set.  The current setup is quite prone to leaks.