Fix Detached notify not getting through
This was due to a longjmp being done immediately after debugger
detachment when the engine is run with the '--debug' option.  Since the
engine immediately terminates in this case, the socket is forcibly
closed before the last message can be sent out.

There is a slight issue with this: If the remote debugger fails to close
its socket for whatever reason, the engine will not terminate.  I'll
have to fix this at some point before 3.0 release.