Fix Eval object inspection
dvalue_dup() doesn't work because InspectHeapObject is picky and expects
a heapptr.  I also took the opportunity to move the dvalue-printing
logic into dvalue.c.  The only thing that stayed in session code was the
object inspection, because dvalue.c has no concept of the
message-passing layer of the debug protocol.