Fix a couple more leaks in minisphere and SSJ
File handles were being leaked in SSJ's source code loader (which may
explain why sometimes source was failing to load, come to think of
it...).

minisphere was leaking the map update and render scripts.  It's
apparently leaking the default map and person scripts as well, but for
the life of me I can't understand how. I checked the code, they're
clearly being freed on shutdown.