Fix SSJ using inordinate amounts of CPU
Turns out setting the Dyad timeout to zero isn't such a great idea in a
console app, it just ends up spinning the CPU waiting for an incoming
message.  It's not an issue in minisphere because the engine already has
its own means of keeping its CPU utilization in check.