Write an actual man page for SSJ
Now includes a detailed overview of the debugger in the description,
command-line arguments and their meanings, and a partial list of
debugger commands.  I would like to complete this before releasing
minisphere 3.0b2.  Seeing as the debugger is the flagship feature of the
minisphere 3.0 release, it needs to be very well documented!