Fix 'ssj --help' output
Due to an extra space on the end of a line, there was an ugly blank line
being printed on 80-character terminals.

Also I also made SSJ's DBZ quote more explicit, because why not?  It was
already pretty obvious where it came from, what with the Cell quotes and
all... ;-)