Add source cache functionality
Technically this is an unnecessary memory hog - it would be enough to
simply do a reverse lookup on the source map to find the file and load
it on-demand.  SSJ will already cache the source locally to avoid
duplicate requests.

Scripts embedded in maps *will* still need to be cached in their
entirety.  That's not implemented yet, though.