Implement SSJ source code download
SSJ now downloads source code from minisphere instead of trying to
locate the source tree itself.  This ensures games can always be
debugged even if the original project is lost (or changed).  As a nice
bonus, this lets SSJ show source for system scripts without the need to
parse the SphereFS path.