SSJ: Auto-select first JS stack frame on pause
This is more friendly, as if a native API function throws an error on
behalf of the game--for example, Abort()--the debugger will stop with
that function on top of the callstack.  For debugging purposes, it's
always more useful to examine the non-native call below that, so SSJ
does the switch automatically.