minisphere 3.0rc5
A ton of minor fixes in this RC, and one rather major one that really
shouldn't have been done during code freeze, but oh well.  Did I get all
the bugs this time?  If not I'm sending a pig after past-me because that
guy clearly sucks and needs to get eaten.