Add frameRate property to `threads`
Since miniRT commandeers the general and map engine frame rates and only
works properly while they remain in sync, it's not really feasible to
allow the legacy functions to be used.  So as a concession, it's now
possible to change the framerate by setting `threads.frameRate`.

Also fixed a few crashy bugs in miniRT/music caused by missed `this`
references during the CommonJS conversion.