minisphere 3.0.0
minisphere 3.0 is a massive release, the biggest in the history of the
project.  It brings several new features, including a full-featured
command-line debugger called SSJ, TypeScript support, and
fully-documented CommonJS modules for miniRT, as well as a slew of minor
fixes and improvements.  It also marks the first time minisphere
binaries have been offered for Linux.