Implement SetLayerWidth/Height()
Either I missed these in my original review of the Sphere API, or I just
didn't see the use at the time.  In any case, with this change it's now
possible to resize map layers at runtime.  As with other changes to map
structures, the change only lasts until the next ChangeMap().