Fix massive Galileo memory leak
This was a HUGE memory leak in the shape object - freeing a shape didn't
free its list of vertices, causing the engine to quickly run out of RAM
when a lot of shapes are generated - which happens, for example, on maps
with a lot of animated tiles.  Kefka's Revenge hit 3 GB within seconds!