Fix tile animation bugs
There were a few things in play here: First of all there was a race
condition regarding the copy-on-write system I had for the tileset's
render texture.  If the map were calculated before a tile could be
animated, it would get the immutable tile atlas rather than the animated
texture.  Second, it was swapping tiles in the texture instead of
overwriting them.  As pretty much all useful tile animations are
looping, this worked out in practice, but it was still pretty sloppy.