Optimize person Ignore List handling
In case of a large ignore list, the overall time cost was around O(n*m)
where `n` is the number of persons and `m` is the length of the ignore
list.  And each entry requires a string comparison, so this became very
expensive very fast.

I was able to optimize the check to O(m) on average by performing the
much cheaper bounding box test first.  Only if an intersection is
detected do we go on to check the ignore list.