Improve compatibility with Duktape v2
I want to make the migration process as painless as possible once
Duktape 2.0 is released.

It seems a lot of the current incompatibilities are due to very early
minisphere code which was written before I had a handle on how
everything in Duktape worked, so it's good that I'm revisiting these
sections now.