Add minisphere 3.3 release notes
Originally the release notes were only for changes with some potential
to break compatibility, but I think it'd be good to start highlighting
important feature additions too.

In other words, it's still a list of breaking changes, but "breaking" is
also defined in terms of impact on developer workflow as well as source
compatibility.