Initialize API based on manifest version
In certain cases having the Pegasus API exposed may cause issues when
running old Sphere games due to name clashes.  Fixed by selectively
enabling the correct API based on the version number in the manifest.

I also made it so that the main script is run as a CommonJS module with
version == 2.  This ensures require() works as expected.