Rename system.doEvents()
I was never very happy with the "doEvents" name, but used it originally
because 1) It matches the name of the method used for this purpose in,
e.g., .NET; and 2) A global function named RunEngine() or similar seemed
even more awkward.  system.run() is nice and concise and communicates
its purpose well, I think.