Implement Joystick#name
This lets you get the name of a joystick as reported by Allegro.  While
it's probably not a good idea to rely on specific joystick names in a
Link query (since they are likely to vary between operating systems),
it's potentially useful for settings screens to show the user which
device is mapped.