Rename `ssj` global to `SSJ`
I originally lowercased `ssj` because it was supposed to be an concrete
object representing the attached debugger.  However, since there won't
always be a debugger attached, the metaphor is a bit broken.  It makes
things clearer if `SSJ` is a namespace.  As such, it should be uppercase
to match typical JS naming conventions.