Fix UTF-8 string validation bug
This was a nasty one.  UTF-8 validation was set up to go until it sees a
NUL terminator, which would be fine--except that `lstr_from_cp1252`
takes a length argument and the buffer passed in may not contain a
terminator.  As a result the function sometimes concluded that valid
UTF-8 text was invalid, and then proceeded to double-encode it.  This
led to all sorts of weird issues elsewhere in the engine.