Improve Sphere v1 compatibility
I resisted this for a long time, but now that Sphere v2 is the preferred
API, I feel this is for the best.  Boolean-typed arguments in Sphere 1.x
are coerced; minisphere was historically strict about this and would
reject anything other than true or false.  Now they are coerced, like
Sphere 1.x--but only for v1 APIs.  v2 APIs will continue to use the
stricter check.