Further improve Sphere 1.x compatibility
There is some Sphere 1.x code in the wild that passes numbers as
strings, expecting coercion.  Supporting this for map layers proved to
be an interesting challenge, since named layers still have to be
supported.  The trick was to do a named layer lookup *first*, and only
if that fails to fall back on type coercion.