Rename JobToken#cancel() to Dispatch.cancel()
I'm think the rule of thumb I will use for the Sphere v2 API is that if
an object doesn't have a constructor, it doesn't get direct methods
either.  There will be exceptions of course (e.g., FileStream), but in
general anything without a constructor is a token, not a proper
autonomous object.