Refactor the Link API a bit for miniRT
Still a lot of work to do here.  Since Link has been around for a while
and is used in production, it has a lot of legacy compatibility baggage.
That's not necessary for Sphere and it'd be better to get rid of the
unnecessary aliasing.  The resulting subset can then be standardized for
Sphere v2.