Implement miniRT "from" module
Link is a bit bloated, so I decided it was best to only have a subset of
its functionality as a standard part of Sphere v2.  Unfortunately I'm
not familiar enough with the Link codebase to feel safe removing stuff.
So instead I implemented a new module based on it.