Remove from#get() query selector
Having thought about this a bit more, I realized the only reason I
considered the "pluck" operation so invaluable is ES5's lack of arrow
functions.  Since TypeScript support is making a comeback, that won't be
an issue soon.  So there's no sense bloating the Sphere API with
convenience functions to compensate.