Improve error blaming algorithms
In the latest build of Duktape it's no longer necessary to defer to
Duktape.act() to get function call metadata as there's a native API to
do it.  Just in case some well-meaning library writer decides to banish
the Duktape object from the global scope, now error handling won't break
as a result.