Overhaul Cell architecture
The original design for Cell dates back to minisphere 2.0 and was never
particularly great to begin with.  This new design hopefully represents
a much cleaner separation of concerns and should be more maintainable
going forward.  Time will tell...

Note that this change has rendered Cell non-functional, as there are no
API bindings in place yet.  I need to come up with a good design
first...