Rename cellscript.c to script.c
Also improved the encapsulation a bit.  It's now set up where the
internals have to create a script_t object which then encapsulates both
a build environment and a Duktape context.  This seems a bit cleaner
than the system used in minisphere where the duk_context pointer is a
global variable.

Come to think of it, it might be possible to get rid of that global
variable now.  I don't think it gets referenced much these days.
Something else to check into...