Refactor Cell some more
A lot of reorganization is going to be needed to get everything in order
for the overhaul, and it's been very difficult to come up with a decent
separation of concerns.  The dynamic between C and JavaScript is quite
different in Cell compared to minisphere; decoupling them has proved to
be quite a challenge.

That said, I think I've finally hit my stride.  Next order of business
from here is to get a Cellscript running again.