Implement CommonJS support in Cell
The thing that frustrates me here is that this duplicates a good chunk
of code from the minisphere codebase.  It would be nice if I could
refactor it so the module loader was shared, but that means the SphereFS
code would *also* have to be shared, and right now that's not possible
due to some pretty substantial differences (in particular, minisphere's
implementation supports SPK with write virtualization, which Cell
doesn't have the architecture to support).  Alas...