Cell: Combine script.c and build.c
So far I haven't come up with a clean way to split the build engine from
the Cellscript executor, as they are pretty tightly linked by nature--a
Tool wraps a JS function, for example.  So until a better solution is
found, it's less confusing to have them combined into a single
component.