Fix crash on startup with invalid JSON
This regression was caused by the change I made a while back to remember
the fullscreen mode when switching games.  If the manifest failed to
parse, minisphere still tried to restore the fullscreen status--except
no screen object has been yet so it ends up crashing the engine.

Also tweaked the output a bit when an uncaught exception occurs.