Rebrand minisphere to miniSphere
The engine is still very lightweight considering everything it does, but
that's no longer the project's main priority, and honestly hasn't been
for quite a while.  The all-lowercase name was starting to seem out of
place.

As a nice bonus, "miniSphere" is camel-cased like the name of a
JavaScript function, which is quite fitting now that it's on the cutting
edge of JavaScript technology.  Right now, I think I might have the only
JS platform championing ES6 modules as a first-class part of the
workflow!