Rename Sphere.reset() -> restart()
"Reset" implies something more passive than what this API actually does.
I originally chose the name Reset because it was supposed to be
analogous to a game console's Reset button.  I need to try to resist the
urge to be clever with naming if I want the API to remain
self-documenting.