Fix Sphere v1 Font#drawZoomedText()
The font's alpha channel was lost due to not setting the blend mode
properly before rendering text.  Honestly the way this is implemented is
pretty inefficient, but I couldn't come up with a better way to do it.
Probably not worth the trouble at this point since Sphere v1 is
deprecated anyway.