Fix Dispatch API job sorting
Each time an update or render job was added the *entire job list* was
resorted, either ascending or descending as appropriate.  Since all
recurring jobs are placed in the same queue but have different ordering
requirements, sorting the entire list one way or the other is incorrect.

As a nice bonus, this fix allowed me to get rid of the duplicate
comparer function.