Re-fix render job priority
The last fix turned out not to work in some cases because the comparison
function was inconsistent.  This fix takes a different
approach--negating the priority.  Since priority is only meaningful
within the same job class, this has the effect of inverting the priority
of render jobs without any detrimental side effects.