Add Sphere v2 Sample class
This serves a similar purpose as the Sphere 1.x SoundEffect API,
allowing sounds to be loaded into memory rather than streamed.  It's
pretty buggy at present, and Sample#play() is a massive memory leak, but
it seems to work at least.

Once this is done I want to also implement the Sphere v1 SoundEffect
API.  Despite the v1 API being deprecated, SoundEffect turns out to be a
Sphere 1.5 feature, so for the sake of migration it should be
implemented.