Fix update/render order
If one quickly glances at the Sphere 1.x MapEngine() loop, it seems like
the render script should be called first, then the update script.  But
thanks to a quirk of Sphere's frame-skipping logic, the first frame
always gets skipped and thus, in practice, the update script is called
first.  A few old games actually depend on this and break outright if
the render script is ever called first--which is usually the case due to
miniSphere's better frameskipping algorithm.

This fixes the bug by always calling the update script first.