Rename Terminal object to Console
Few, if any, games call their command console a "terminal".  Better to
stick to gamer lingo than try to be clever with the terminology and end
up scaring newbies away.

The main reason I didn't want to do this before is because it presents
some confusion with the built-in console.log().  But I'm thinking now
it's probably best to remove that--printing to the console is useful for
debugging, but it should be namespaced accordingly.  In other words, the
SSJ object is coming back, baby!