Implement Sphere v2 Game class instantiation
If the main module default-exports a constructor or class, the engine
will instantiate it on startup and call its start() method.  This is a
modern analogue to Sphere v1's game() function and is very powerful when
used in tandem with ES6 classes.

Of course, even with this change, it remains completely acceptable to
place startup code at the top level--which is actually *preferable* when
not using a transpiler--since not allowing that would be against the
spirit of JS.