Use FOV/aspect instead of frustum for project3D()
I'm a bit torn on this change, and may revert it.  While FOV/aspect is
much easier to understand than frustum edges, the Core API is meant to
be low-level and defines what is possible to do with the engine.  So
this takes some freedom away from game developers, and I'm naturally
hesitant about doing that.