Move Shader code into galileo.c
This was probably the wrong direction to take this in; usually I would
try to split different objects into their own source files.  At least in
the case of Galileo though, all the components are pretty tightly
coupled so it's cleaner overall for them to be in the same source file.

Oh, and Allegro 5.2 is an absolute requirement now.  I've had to
maintain a lot of undertested fallback code to continue supporting
Allegro 5.0, and it's just going to get worse as I use more of the
library.  Best to just bump the requirement at this point.