Rename SoundStream to AudioStream
I found SoundStream to a bit of a confusing name; at times I wanted to
assume it was something returned by, e.g., Sound#play() to represent an
active playback.  Calling it AudioStream instead makes it look less like
a subordinate object of Sound and hopefully makes its purpose clearer in
the process.