Tentatively fix Galileo render-to-surface bug
Due to a discrepency between Allegro and miniSphere vis-a-vis shader
tracking, Galileo was often using the wrong shader when drawing a model
or shape.  The fix for now is to always set the shader, which is
potentially inefficient.  A proper fix would be more involved and
require tracking shaders per texture.