Refactor internal Spriteset object
spriteset.c contained some of the oldest code in the engine.  It was one
of the first parts of the engine I developed and the code has barely
been touched since then.  I figured now was a good time to overhaul it,
as part of implementing Sphere 1.x-compatible spriteset handling.

There might still be bugs in this implementation; more testing will be
needed before this makes it into a release.