Implement Sphere 1.x-compatible spriteset semantics
This allows spriteset objects to be modified directly in JS and then
passed back to SetPersonSpriteset(), as can be done in Sphere 1.x.  That
didn't work before because the property accesses weren't trapped, so
internally the spriteset wasn't affected by JS-side modifications.

The fix was to do what Sphere 1.x apparently does: Reconstruct a new
internal spriteset object from the properties of the object passed in.
That's inefficient, but if it's the only way to ensure full backward
compatibility, then so be it.

Fixes #175.