Implement Sphere v1 CreateSpriteset()
Let me just say that the design of this function is... somewhat
questionable.  It creates a spriteset with blank images, which would be
fine... except images in Sphere are read-only, so the blank images waste
memory.  However, *because* they're read-only, that allowed me to just
reuse the same internal image object in all slots.