Add internal support for indexed shapes
Next step: Exposing this functionality through the API.  I didn't think
this was going to make it into miniSphere 4.7, but it turns out it was
rather easy to implement.

What's not supported yet: Reusable and hot-swappable index and vertex
buffers.  Both are maintained internally by the Shape object with no
possibility of reuse.  That requires some refactoring of the Galileo API
that I'm not ready to make yet.