Work around Allegro clipping bug
It turns out that the GL_INVALID_VALUE error I had to clear was caused
by Allegro calling glScissor() with a negative size when the clipping
box is out of bounds.  This also meant that the scissor box wasn't
actually set, so I have no choice but to add a workaround.