Clean up after Map Engine/Persons merger
I realized while doing this that a lot of the necessary hooks for
accessing the map engine from outside were already in place, I just
didn't have prototypes for them in the header.  This is an advantage of
the current naming convention, it's easy to see at a glance what's
missing on a per-component basis.