Start refactoring map_engine.c
The API layer should consist almost entirely of stubs, with the real
code to implement the functionality elsewhere.  In large part, this is
already the case--except for the map engine.  Many of the API stubs
access internal map engine variables directly, which makes it impossible
to separate them from the map engine proper.  So it seems a nice
refactoring session is in order here...