Fix SphereFS canonicalization of startup script
The startup script filename wasn't properly canonicalized in Sphere v2
games; as stupidity would have it, the handling for the `$/` prefix
relied on the broken behavior, causing it not to work in all cases.  The
fix is to always canonicalize the script filename and NOT automatically
add a `@/` prefix in dollar-path handling.

Fixes #205.