Refactor WindowStyle implementation
Another bit of code that was hastily thrown together early in the
project's life and then barely touched afterwards, this was long overdue
for an update.

I also decided to start a new internal convention: reference counted
objects use "unref" instead of "free" for refcounted objects.  It wasn't
always clear when seeing `whatever_free()` whether the object was
refcounted or not, so this makes it explicit.