Rename all refcounted *_free() to *_unref()
This makes it clear at the time an object is unref'd that the call may
or may not free it.  Before "free" was used for both refcounted and
non-refcounted objects, leading to a few bugs in the past where objects
got freed prematurely thanks to me not realizing they weren't reference
counted.