Remove trailing spaces from C sources
Visual Studio has a very bad habit of adding phantom tabs at the
beginning of blank lines to match the indent level, *despite* supporting
virtual space in the IDE itself up to the surrounding indent level.
I've known about the problem for a while, but have been too lazy to fix
all the sources, as per usual.

I also documented the specific behavior of FS.createDirectory().  I can
say from experience that it's never really obvious what happens when
attempting to create a directory that already exists, so I thought it
was best to explicitly lay out the behavior in the documentation.