Implement DirectoryStream in Cell
So it seems that, when I get close to a release deadline, I just become
really efficient.  Not to say I'm a slouch any other time, but I just
start implementing things left and right this close to a release.  This
is how miniSphere got put together so quickly in the first place, I
suppose.

Fixes #209.