Add DirectoryStream#dispose()
JS is garbage collected, but there's no guarantee of when that happens,
so it's best to let the user close the directory.  Currently it's not
necessary since the whole thing is read in one go, but since a future
optimization might involve holding the directory open, best to make it
part of the API now.