Fix crash on Shape#draw() with no arguments
This regression is connected with the change to backbuffer handling in
mS 4.6.  In the old system passing NULL as the surface parameter meant
draw to the screen, however now that the backbuffer is a real image_t
object, that's no longer valid.  Shape#draw() apparently didn't get the
memo. :o)