Add direct throw mechanism to JSAL
This allows native calls to bail out on the spot by throwing a JS
exception.  I *think* this will even work if we're several native calls
deep and the topmost one throws an exception.

While this is somewhat awkward--and against my better judgment--it's
necessary unless I want to spend a week refactoring the rest of
miniSphere.  Besides the fact that deferred throws can't be emulated *at
all* under Duktape's longjmp() model.