Don't burn the CPU while waiting for debugger
I wanted to keep the sockets code generic but still enable blocking
reads, so the only way to do that without spinning the CPU nonstop was
to add an idle handler.

If no idle handler is set, then Dyad is initialized with a nonzero
timeout, otherwise SSj runs the CPU at full tilt too.