Fix sounds not resuming after SSj breakpoint
The audio engine frees sound references once they stop playing.  The
trouble in this case is that "paused" also counts as "not playing".
Normally that doesn't cause any bugs: the mechanism exists to prevent
sounds from suddenly stopping after a Sound object goes out of scope,
and if it's out of scope it can't be paused.  Of course, now that the
*debugger* pauses them... well, you can see the problem here.