Add FS.evaluateScript() for use with browser code
Browsers don't yet widely support modules, so scripts written for the
Web are necessarily designed to run in the global scope.  So far the
only way to load such scripts was to call Sphere v1 EvaluateScript(),
which besides being deprecated, is not SphereFS compliant.

Since loading browser code is still a valid use case, I figured it was
high time I brought the ability to pure Sphere v2 code.