Rename Color#fade() to Color#fadeTo()
It might seem like a waste of time to backport API changes, but I'm
trying to minimize the disruption as much as possible.  5.0 will still
bring a lot of breaking changes to the runtime, but if I can keep the
Core API in sync at least, the upgrade shouldn't be as painful.