Refactor JSAL mJS module handling
I originally hacked around the asynchronicity of ChakraCore's module
loading APIs, but that made it impossible to support import(), which is
promise-based.  This change allows modules to be loaded and executed
from within the event loop.

There are still a few glitches related to exception handling, notably
that an error thrown by import() will crash the entire game rather than
just rejecting the promise.  That will need to be fixed before the next
beta.