Avoid deferring chained Dispatch jobs
It turns out that, because of the way promises work, `await` always
defers to the event loop even if the operation completes immediately.
That avoids releasing Zalgo, but made it impractical in many cases to
use `await` at all.  This fixes it by removing the "next tick"
enforcement for chained jobs.