Make Sphere.sleep() awaitable
I'd like to move away from any kind of blocking primitives in the API
and go to a completely async design, as in Node.js.  This is a little
bittersweet as it feels like the end of an era.  No more synchronous
logic...  Oh well.  There's always the Sphere v1 API which will probably
remain supported forever.  JavaScript's backward compatibility pretty
much dates back to its inception, so why not Sphere? :o)