Add warning about 'await' for Thread#on_render()
A render handler should never await: the handler will just be called
again on the next frame and that will almost assuredly cause odd things
to happen, if only because it will violate the developer's assumptions.
The ramifications are difficult to explain adequately in concise
documentation, so we'll just say it's undefined behavior and leave it at
that.