Clean up CHANGELOG.md a bit
The release is getting closer, might as well get as much tedious cleanup
out of the way as I can.  I still have bugs to fix and I'd prefer not to
have to do all the boring stuff at the last minute like I usually end up
having to do.