Fix generic error for module-not-found
The generic error thrown by ChakraCore when a module import fails
doesn't include the specifier; getting it to use a custom message
counterintuitively means actually creating a module record for the
nonexistent module.

The real question we need to ask is why the modules stopped existing in
the first place.  For what it's worth?  I blame this thing -> :pig2: